requirements.txt
aiofiles~=23.1.0
starlette~=0.27.0
SQLAlchemy~=2.0.15
fastapi~=0.95.2
pydub~=0.25.1
pydantic~=1.10.7
uvicorn~=0.22.0
aiohttp~=3.8.4
loguru~=0.7.0
python-dotenv~=1.0.0
alembic~=1.11.1